REF: Move `octue.compatibility`
